+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   - route_target_import, route_target_import_evpn
   - route_target_export, route_target_export_evpn
 - Extended cisco_vrf with `vni`
+- Extended cisco_bgp with `fast_external_fallover`, `flush_routes`, `isolate`
 
 ### Removed